pyproject.toml
[tool.poetry]
name = "my_hackathon_bot"
version = "0.1.0"
description = ""
authors = ["WebDev22"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.15"
pymongo = "^4.3.3"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
